♻️ F08: ORDER > Main > Added MenuEmpty comp
